backend/requirements.txt
duckdb
pyspark
deltalake